requirements.txt
Django==2.2.3
django-reversion==3.0.4
pytz==2019.1
sqlparse==0.3.0
